Fix broken test of "EqualTo"
The test that this commit will fix tests whether the casting in
"EqualTo" works when it compares DateTime objects. However, it uses
"now" which will always have different values every time it gets parsed.

This commit will replace "now" with a fixed date to make the tests pass
and still making the test valid.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>